+++ src/main/java/org/example/verestrotask/client/account/dto/AccountResponseDto.java
@@ -1,6 +1,7 @@
 package org.example.verestrotask.client.account.dto;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
-public record AccountResponseDto(int identifier, BigDecimal balance, int dayLimit) {
+public record AccountResponseDto(BigInteger identifier, BigDecimal balance, int dayLimit) {
 }
