+++ src/main/java/org/example/verestrotask/client/account/dto/Transfer.java
@@ -1,6 +1,7 @@
 package org.example.verestrotask.client.account.dto;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
-public record Transfer(int identifier, BigDecimal amount) {
+public record Transfer(BigInteger identifier, BigDecimal amount) {
 }
